almost done sysinfo

+++ kernel/proc.c
@@ -696,8 +696,10 @@ procdump(void)
   }
 }
 
-uint32 count_active_proc(void) {
-  uint count = 0;
+uint64
+count_active_proc(void)
+{
+  uint64 count = 0;
   struct proc *p;
   
   for (p = proc; p < &proc[NPROC]; ++p) // contigous
